requirements.txt
asgiref==3.6.0
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
distro==1.8.0
Django==4.1.5
django-dotenv==1.4.2
django-filter==22.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==3.2.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
mysqlclient==2.1.1
packaging==22.0
paramiko==2.12.0
Pillow==9.4.0
protobuf==3.20.1
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyrsistent==0.19.3
pytz==2022.7
PyYAML==5.4.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
texttable==1.6.7
uritemplate==4.1.1
urllib3==1.26.13
websocket-client==0.59.0
